+++ weni/broadcasts/tests/test_sender.py
@@ -28,6 +28,7 @@ def create_context(
         globals=globals or {},
         contact=contact or {},
         project=project or {},
+        constants={},
     )